requirements.txt
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
Django==2.2.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-modelcluster==4.4
django-recaptcha==2.0.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.5
html5lib==1.0.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
Pygments==2.4.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.3
wagtail==2.5.1
wagtail-django-recaptcha==1.0
wagtail-modelchooser==2.2.1
-e git+https://github.com/djdmorrison/wagtail-orderable.git@76fc505cdc7ed227fcd91abc1469e3e134d21940#egg=wagtail_orderable
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
